♻️🎪 Extract animations email service
